Split up config files between consumer and producer
